requirements.txt
discord
python-dotenv
requests
bs4